requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-extra-fields==1.2.4
django-rest-framework==0.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
Pillow==6.1.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
